Align source code consistently to be easier to follow or print-out.
This massive patch indents_2 the Replica Island source code a bit more
consistently for developers who want to learn the source code and like
the code to be more consistent in indentation. This is a branch of the
original source code at: https://code.google.com/p/replicaisland/
with added patches found elsewhere merged in.

You can diff (use ignore spaces for comparison) between original and
this update - Most edits are in remarks to get them to fit in 80chars,
otherwise everything else is just indentation.